+++ front/lib/resources/data_source_resource.ts
@@ -3,6 +3,7 @@ import type {
   DataSourceType,
   ModelId,
   Result,
+  UserType,
 } from "@dust-tt/types";
 import { Err, formatUserFullName, Ok } from "@dust-tt/types";
 import type {
@@ -224,6 +225,18 @@ export class DataSourceResource extends ResourceWithVault<DataSource> {
     // Update the current instance with the new values to avoid stale data
     Object.assign(this, affectedRows[0].get());
     return [affectedCount];
+  }
+
+  async setEditedBy(auth: Authenticator, user: UserType) {
+    if (!auth.isAdmin()) {
+      throw new Error(
+        "Unexpected call to DataSourceResource.setEditedBy by non admin"
+      );
+    }
+    await this.update({
+      editedByUserId: user.id,
+      editedAt: new Date(),
+    });
   }
 
   private makeEditedBy(
